Properly escape table names
The method CppSQLite3DB::tableExist was not escaping the table names in the
sql query properly, which caused exceptions when the table name contains
quotes.

The fix was simply to change the string formatting to use the sqlite specific
printf variant (through CppSQLite3Buffer) with a %Q. This has the nice side
effect of also removing a warning when building with visual studio.